rebase statistic from main to View and Controller
